feat: rename filtersidebar and nft topbar (#1238)
