+++ webapp/src/components/AssetFilters/PriceFilter/PriceFilter.css
