+++ webapp/src/components/AssetFilters/index.ts
@@ -1,0 +1,3 @@
+import AssetFilters  from './AssetFilters.container'
+
+export { AssetFilters };
