+++ webapp/src/components/AssetTopbar/AssetTopbar.types.ts
@@ -1,9 +1,9 @@
 import { openModal } from 'decentraland-dapps/dist/modules/modal/actions'
-import { AssetType } from '../../../../modules/asset/types'
-import { BrowseOptions } from '../../../../modules/routing/types'
-import { Section } from '../../../../modules/vendor/routing/types'
-import { View } from '../../../../modules/ui/types'
-import { clearFilters } from '../../../../modules/routing/actions'
+import { AssetType } from '../../modules/asset/types'
+import { BrowseOptions } from '../../modules/routing/types'
+import { Section } from '../../modules/vendor/routing/types'
+import { View } from '../../modules/ui/types'
+import { clearFilters } from '../../modules/routing/actions'
 
 export type Props = {
   count: number | undefined
