+++ webapp/src/components/AssetTopbar/SelectedFilters/SelectedFilters.container.ts
@@ -1,14 +1,14 @@
 import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
-import { RootState } from '../../../../../modules/reducer'
-import { browse } from '../../../../../modules/routing/actions'
+import { RootState } from '../../../modules/reducer'
+import { browse } from '../../../modules/routing/actions'
 import {
   getCurrentBrowseOptions,
   getSection
-} from '../../../../../modules/routing/selectors'
+} from '../../../modules/routing/selectors'
 import { MapStateProps, MapDispatchProps } from './SelectedFilters.types'
 import { SelectedFilters } from './SelectedFilters'
-import { isLandSection } from '../../../../../modules/ui/utils'
+import { isLandSection } from '../../../modules/ui/utils'
 
 const mapState = (state: RootState): MapStateProps => {
   return {
