+++ webapp/src/components/AssetTopbar/SelectedFilters/utils.ts
@@ -1,10 +1,10 @@
+import { collectionAPI } from '../../../modules/vendor/decentraland'
 import { GenderFilterOption, WearableGender } from '@dcl/schemas'
-import { collectionAPI } from '../../../../../modules/vendor/decentraland'
 import {
   AVAILABLE_FOR_FEMALE,
   AVAILABLE_FOR_MALE,
   getBodyShapeValue
-} from '../../../../FiltersSidebar/BodyShapeFilter/utils'
+} from '../../AssetFilters/BodyShapeFilter/utils'
 
 export async function getCollectionByAddress(address: string) {
   const { data } = await collectionAPI.fetch({
