+++ webapp/src/components/AssetTopbar/index.ts
@@ -1,0 +1,3 @@
+import AssetTopbar from './AssetTopbar.container'
+
+export default AssetTopbar
