--- webapp/src/components/FiltersSidebar/index.ts
@@ -1,3 +1,0 @@
-import FiltersSidebar  from './FiltersSidebar.container'
-
-export { FiltersSidebar };
